remove unnecessary HomeScreen file
